a2dp: Set the default A2DP bitrate to high quality
Allows bluedroid to calculate the proper A2DP bitpool size.
This resolves the main A2DP audio quality issues with Android 4.2,
but does not address the difference in audio output gain/volume as
compared with Android 4.1.

Setting the output volume level to approximately 80% in conjunction
with this change seems to result in comparable audio quality

Change-Id: Ic05c3278ebe9288d00926fab8b7c1a646fd861cd